refactor add recipe into separate file
