+++ food_db/food_db_app/views/__init__.py
@@ -1,3 +1,4 @@
 from .apis import *
 from .pages import *
 from .utils import *
+from .add_edit_recipe import add_recipe
